refactors backoffconfig to namer core
