+++ namerd/storage/consul/src/main/scala/io/buoyant/namerd/storage/consul/ConsulDtabStoreInitializer.scala
@@ -1,12 +1,15 @@
 package io.buoyant.namerd.storage.consul
 
 import com.fasterxml.jackson.annotation.JsonIgnore
+import com.twitter.finagle.service.Backoff
 import com.twitter.finagle.tracing.NullTracer
 import com.twitter.finagle.{Http, Path}
 import io.buoyant.config.types.Port
 import io.buoyant.consul.utils.RichConsulClient
 import io.buoyant.consul.v1.{ConsistencyMode, KvApi}
-import io.buoyant.namerd.{BackoffConfig, DtabStore, DtabStoreConfig, DtabStoreInitializer}
+import io.buoyant.namer.BackoffConfig
+import com.twitter.conversions.time._
+import io.buoyant.namerd.{DtabStore, DtabStoreConfig, DtabStoreInitializer}
 
 case class ConsulConfig(
   host: Option[String],
@@ -17,7 +20,7 @@ case class ConsulConfig(
   readConsistencyMode: Option[ConsistencyMode] = None,
   writeConsistencyMode: Option[ConsistencyMode] = None,
   failFast: Option[Boolean] = None,
-  backoffDuration: Option[BackoffConfig] = None
+  backoff: Option[BackoffConfig] = None
 ) extends DtabStoreConfig {
   import ConsulConfig._
 
@@ -25,7 +28,7 @@ case class ConsulConfig(
   override def mkDtabStore: DtabStore = {
     val serviceHost = host.getOrElse(DefaultHost)
     val servicePort = port.getOrElse(DefaultPort).port
-    val maxBackoffMs = backoffDuration.getOrElse(DefaultMaxDurationMs).mk
+    val backoffs = backoff.map(_.mk).getOrElse(DefaultBackoff)
 
     val service = Http.client
       .interceptInterrupts
@@ -35,7 +38,7 @@ case class ConsulConfig(
       .withTracer(NullTracer)
       .newService(s"/$$/inet/$serviceHost/$servicePort")
     new ConsulDtabStore(
-      KvApi(service, maxBackoffMs),
+      KvApi(service, backoffs),
       pathPrefix.getOrElse(Path.read("/namerd/dtabs")),
       datacenter = datacenter,
       readConsistency = readConsistencyMode,
@@ -47,7 +50,7 @@ case class ConsulConfig(
 object ConsulConfig {
   val DefaultHost = "localhost"
   val DefaultPort = Port(8500)
-  val DefaultMaxDurationMs = BackoffConfig(Some(1), Some(10000))
+  val DefaultBackoff = Backoff.decorrelatedJittered(1.millis, 1.minute)
 }
 
 class ConsulDtabStoreInitializer extends DtabStoreInitializer {
